requirements.txt
Django==5.0.1
Pillow==10.4.0
python_dateutil==2.8.2
crispy-bootstrap4==2024.1
django-crispy-forms==2.1

